bug: handle memtable pointers properly
The random failure of _tidesdb_new_column_family is happen
to be due to memtable pointers wasn't zeroed that lead to
containing random garbage for them

Next: memtable check for NULL pointer is failing for any
of the configuration (i.e. TDB_MEMTABLE_SKIP_LIST or TDB_MEMTABLE_HASH_TABLE)
however we actually want ONE of them to contain a valid pointer instead.